pyproject.toml
[tool.poetry]
name = "directory-info"
version = "0.1.0"
description = ""
authors = ["Ivan Timofeev <timofeevis89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
